Reorganised unit tests locations - moved all of them to a dedicated folder
